+++ mainAspx/ClassCheck_Modify.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="admin1-1-1.aspx.cs" Inherits="admin1_1_1" %>
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="ClassCheck_Modify.aspx.cs" Inherits="admin1_1_1" %>
 
 <!DOCTYPE html>
