+++ mainAspx/ProjectDo.aspx.cs
@@ -7,21 +7,24 @@ using System.Web.UI.WebControls;
 using System.Data;
 using System.Data.SqlClient;
 
-public partial class admin2_2 : System.Web.UI.Page
+public partial class admin2_3 : System.Web.UI.Page
 {
     protected void Page_Load(object sender, EventArgs e)
     {
-        Database mDB = new Database();
-        string sql = "select [工号],[教师姓名],[提交时间],[状态] from [授导学生表]";
+        Database mDB = new Database();        
+        string sql = "select [工号],[教师姓名],[提交时间],[状态] from [项目实践表]";
+        //获取选择的项目信息
+        mDB.ExecuteSQL(sql);
         GridView1.DataSource = mDB.getDataTable(sql);
         GridView1.DataBind();
+        //显示信息
         mDB.Close();
     }
     protected void GridView1_SelectedIndexChanged(object sender, EventArgs e)
     {
-        Session["TeacherID4"] = GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text.ToString();
-        Session["TeacherName4"] = GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text.ToString();
-        Response.Redirect("admin2-2-1-1.aspx");
+        Session["TeacherID5"] = GridView1.Rows[GridView1.SelectedIndex].Cells[1].Text.ToString();
+        Session["TeacherName5"] = GridView1.Rows[GridView1.SelectedIndex].Cells[2].Text.ToString();
+        Response.Redirect("ProjectDo_Modify.aspx");
     }
     protected void TextBox1_TextChanged(object sender, EventArgs e)
     {
@@ -30,7 +33,9 @@ public partial class admin2_2 : System.Web.UI.Page
     protected void Button1_Click(object sender, EventArgs e)
     {
         Database mDB = new Database();
-        string sql = "select [工号],[教师姓名],[提交时间],[状态] from [授导学生表] where [工号]+[教师姓名] like'%" + TextBox1.Text + "%'";
+        string sql = "select [工号],[教师姓名],[提交时间],[状态] ";
+        sql += "from [项目实践表] where [工号]+[教师姓名] like'%" + TextBox1.Text + "%'";
+        //模糊搜索
         GridView1.DataSource = mDB.getDataTable(sql);
         GridView1.DataBind();
         mDB.Close();
